# Improve performance of repeating strings