This commit improves the performance of repeating strings, by copying
the result string instead of the string being repeated. Also it adds
an error message when the result string is too long.

```sh
 $ env q='"abcde" * 10000000 | length' hyperfine --warmup 10 './jq-old -n "$q"' './jq-new -n "$q"'
Benchmark 1: ./jq-old -n "$q"
  Time (mean ± σ):     375.4 ms ±   2.8 ms    [User: 363.7 ms, System: 10.5 ms]
  Range (min … max):   371.8 ms … 380.5 ms    10 runs
 
Benchmark 2: ./jq-new -n "$q"
  Time (mean ± σ):     115.9 ms ±   0.7 ms    [User: 107.7 ms, System: 7.1 ms]
  Range (min … max):   114.8 ms … 117.2 ms    24 runs
 
Summary
  ./jq-new -n "$q" ran
    3.24 ± 0.03 times faster than ./jq-old -n "$q"

 $ env q='"a" * 50000000 | length' hyperfine --warmup 10 './jq-old -n "$q"' './jq-new -n "$q"'
Benchmark 1: ./jq-old -n "$q"
  Time (mean ± σ):     813.9 ms ±  32.9 ms    [User: 779.1 ms, System: 15.2 ms]
  Range (min … max):   782.9 ms … 886.0 ms    10 runs
 
Benchmark 2: ./jq-new -n "$q"
  Time (mean ± σ):     116.4 ms ±   1.7 ms    [User: 107.8 ms, System: 7.3 ms]
  Range (min … max):   114.8 ms … 120.5 ms    25 runs
 
Summary
  ./jq-new -n "$q" ran
    6.99 ± 0.30 times faster than ./jq-old -n "$q"
```